+++ client/helpers/handlebars.js
@@ -15,8 +15,8 @@ UI.registerHelper('eachWithRank', function(items, options) {
 });
 
 UI.registerHelper('getSetting', function(setting, defaultArgument){
-  var defaultArgument = (typeof defaultArgument !== 'undefined') ? defaultArgument : '';
-  var setting = getSetting(setting, defaultArgument);
+  defaultArgument = defaultArgument || '';
+  setting = getSetting(setting, defaultArgument);
   return setting;
 });
 UI.registerHelper('isLoggedIn', function() {
@@ -32,13 +32,13 @@ UI.registerHelper('canComment', function() {
   return can.comment(Meteor.user());
 });
 UI.registerHelper('isAdmin', function(showError) {
-  if(isAdmin(Meteor.user())){
+  if (isAdmin(Meteor.user())) {
     return true;
-  }else{
-    if((typeof showError === "string") && (showError === "true"))
-      flashMessage(i18n.t('sorry_you_do_not_have_access_to_this_page'), "error");
-    return false;
   }
+  if ((typeof showError === 'string') && (showError === 'true')) {
+    flashMessage(i18n.t('sorry_you_do_not_have_access_to_this_page'), 'error');
+  }
+  return false;
 });
 UI.registerHelper('canEdit', function(item) {
   return can.edit(Meteor.user(), item, false);
@@ -48,19 +48,19 @@ UI.registerHelper('log', function(context){
   console.log(context);
 });
 
-UI.registerHelper("formatDate", function(datetime, format) {
+UI.registerHelper('formatDate', function(datetime, format) {
   Session.get('momentLocale'); // depend on session variable to reactively rerun the helper
   return moment(datetime).format(format);
 });
 
-UI.registerHelper("timeAgo", function(datetime) {
+UI.registerHelper('timeAgo', function(datetime) {
   Session.get('momentLocale'); // depend on session variable to reactively rerun the helper
-  return moment(datetime).fromNow()
+  return moment(datetime).fromNow();
 });
 
-UI.registerHelper("sanitize", function(content) {
-  console.log('cleaning up…')
-  console.log(content)
+UI.registerHelper('sanitize', function(content) {
+  console.log('cleaning up…');
+  console.log(content);
   return cleanUp(content);
 });
 
@@ -69,14 +69,16 @@ UI.registerHelper('pluralize', function(count, string) {
   return i18n.t(string);
 });
 
-UI.registerHelper("profileUrl", function(userOrUserId) {
-  var user = (typeof userOrUserId === "string") ? Meteor.users.findOne(userOrUserId) :  userOrUserId;
-  if (!!user)
+UI.registerHelper('profileUrl', function(userOrUserId) {
+  var user = (typeof userOrUserId === 'string') ? Meteor.users.findOne(userOrUserId) :  userOrUserId;
+  if (!!user) {
     return getProfileUrl(user);
+  }
 });
 
-UI.registerHelper("userName", function(userOrUserId) {
-  var user = (typeof userOrUserId === "string") ? Meteor.users.findOne(userOrUserId) :  userOrUserId;
-  if (!!user)
+UI.registerHelper('userName', function(userOrUserId) {
+  var user = (typeof userOrUserId === 'string') ? Meteor.users.findOne(userOrUserId) :  userOrUserId;
+  if (!!user) {
     return getUserName(user);
+  }
 });