+++ packages/telescope-daily/lib/routes.js
@@ -8,6 +8,7 @@ var coreSubscriptions = new SubsManager({
 PostsDailyController = RouteController.extend({
   
   template: function() {
+    // use a function to make sure the template is evaluated *after* any template overrides
     return getTemplate('postsDaily');
   },