+++ packages/telescope-embedly/lib/embedly.js
@@ -26,7 +26,6 @@ var mediaProperty = {
 }
 addToPostSchema.push(mediaProperty);
 
-
 postThumbnail.push({
   template: 'postThumbnail', 
   order: 15
@@ -68,3 +67,9 @@ var thumbnailHeightProperty = {
   }
 }
 addToSettingsSchema.push(thumbnailHeightProperty);
+
+// add callback that adds "has-thumbnail" or "no-thumbnail" CSS classes
+postClassCallbacks.push(function (post, postClass){
+  var thumbnailClass = !!post.thumbnailUrl ? "has-thumbnail" : "no-thumbnail";
+  return postClass + " " + thumbnailClass;
+});