+++ packages/telescope-newsletter/lib/client/templates/newsletter_banner.js
@@ -71,7 +71,7 @@ Meteor.startup(function () {
           $banner.removeClass('show-loader');
           if(error){
             console.log(error);
-            flashMessage(error.message, "error");
+            flashMessage(error.reason, "error");
           }else{
             console.log(result);
             confirmSubscription();