+++ packages/telescope-subscribe-to-posts/lib/server/publications.js
@@ -1,0 +1,5 @@
+Meteor.publish('userSubscribedPosts', function(terms) {
+  var parameters = getPostsParameters(terms);
+  var posts = Posts.find(parameters.find, parameters.options);
+  return posts;
+});