+++ History.md
@@ -1,3 +1,14 @@
+## v0.14.0 “GridScope”
+
+* Adeed Grid Layout option.
+* Cleaned up vote click handling functions and added tests (thanks @anthonymayer!).
+* Added `threadModules` zone.
+* Added `upvoteCallbacks` and `downvoteCallbacks` callback arrays.
+* Fix “post awaiting moderation” message bug.
+* Added `postApproveCallbacks` callback array.
+* Added notifications for pending and approved posts, for admins and end users.
+* Renaming "digest" view to "singleday".
+
 ## v0.13.0 “ComponentScope”
 
 * Tweaked comments layout in Hubbble theme.